Got syscalls working

+++ kernel/c/tss.c
@@ -4,7 +4,7 @@
 
 #include "../h/tss.h"
 
-void tss_init(uintptr_t kernel_stack){
+void tss_init(uint64_t kernel_stack){
     tss_load();
     memset((void *) tss, 0, sizeof(tss_entry_t));
     tss->rsp0 = kernel_stack;
